domain is_locked

+++ rust/src/gen/base/domain/domain_service.rs
@@ -119,6 +119,17 @@ pub async fn update_by_id<'a>(
   options: Option<Options>,
 ) -> Result<String> {
   
+  let is_locked = domain_dao::get_is_locked_by_id(
+    ctx,
+    id.clone(),
+    None,
+  ).await?;
+  
+  if is_locked {
+    let err_msg = i18n_dao::ns(ctx, "不能修改已经锁定的数据".to_owned(), None).await?;
+    return Err(SrvErr::msg(err_msg).into());
+  }
+  
   let res = domain_dao::update_by_id(
     ctx,
     id,
@@ -137,6 +148,28 @@ pub async fn delete_by_ids<'a>(
   ids: Vec<String>,
   options: Option<Options>,
 ) -> Result<u64> {
+  
+  let len = ids.len();
+  let ids0 = ids.clone();
+  let mut ids: Vec<String> = vec![];
+  for id in ids0 {
+    let is_locked = domain_dao::get_is_locked_by_id(
+      ctx,
+      id.clone(),
+      None,
+    ).await?;
+    
+    if is_locked {
+      continue;
+    }
+    
+    ids.push(id);
+  }
+  if ids.len() == 0 && len > 0 {
+    let err_msg = i18n_dao::ns(ctx, "不能删除已经锁定的数据".to_owned(), None).await?;
+    return Err(SrvErr::msg(err_msg).into());
+  }
+  let ids = ids;
   
   let num = domain_dao::delete_by_ids(
     ctx,
@@ -195,6 +228,46 @@ pub async fn enable_by_ids<'a>(
 ) -> Result<u64> {
   
   let num = domain_dao::enable_by_ids(
+    ctx,
+    ids,
+    is_locked,
+    options,
+  ).await?;
+  
+  Ok(num)
+}
+
+/// 根据 ID 查找是否已锁定
+/// 已锁定的记录不能修改和删除
+/// 记录不存在则返回 false
+#[instrument(skip(ctx))]
+#[allow(dead_code)]
+pub async fn get_is_locked_by_id<'a>(
+  ctx: &mut impl Ctx<'a>,
+  id: String,
+  options: Option<Options>,
+) -> Result<bool> {
+  
+  let is_locked = domain_dao::get_is_locked_by_id(
+    ctx,
+    id,
+    options,
+  ).await?;
+  
+  Ok(is_locked)
+}
+
+/// 根据 ids 锁定或者解锁数据
+#[instrument(skip(ctx))]
+#[allow(dead_code)]
+pub async fn lock_by_ids<'a>(
+  ctx: &mut impl Ctx<'a>,
+  ids: Vec<String>,
+  is_locked: u8,
+  options: Option<Options>,
+) -> Result<u64> {
+  
+  let num = domain_dao::lock_by_ids(
     ctx,
     ids,
     is_locked,
